# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ever Works Website Template API",
    "description": "REST API documentation for the Ever Works website template. This template provides a complete foundation for building modern web applications with authentication, user management, favorites, and content management features.",
    "version": "1.0.0",
    "contact": {
      "name": "Ever Works Template",
      "url": "https://demo.ever.works"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://demo.ever.works",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/admin/categories/{id}": {
      "get": {
        "tags": ["Admin - Categories"],
        "summary": "Get category by ID (Admin)",
        "description": "Returns a specific category by its ID. Data is loaded from Git repository. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the category",
            "example": "productivity-tools"
          }
        ],
        "responses": {
          "200": {
            "description": "Category retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCategoryResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "productivity-tools",
                    "name": "Productivity Tools"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Unauthorized. Admin access required."
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Category not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "put": {
        "tags": ["Admin - Categories"],
        "summary": "Update category (Admin)",
        "description": "Updates a category's name. Category ID cannot be changed. Data is stored in Git repository. Validates name length (2-50 chars) and checks for duplicates. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the category",
            "example": "productivity-tools"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              },
              "example": {
                "name": "Enhanced Productivity Tools"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCategoryResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "productivity-tools",
                    "name": "Enhanced Productivity Tools"
                  },
                  "message": "Category updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "name_too_short": {
                    "summary": "Name too short",
                    "value": {
                      "success": false,
                      "error": "Category name must be at least 2 characters long"
                    }
                  },
                  "name_too_long": {
                    "summary": "Name too long",
                    "value": {
                      "success": false,
                      "error": "Category name must be no more than 50 characters long"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Unauthorized. Admin access required."
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Category with ID \"productivity-tools\" not found"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - category name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Category with name \"Enhanced Productivity Tools\" already exists"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": ["Admin - Categories"],
        "summary": "Delete category (Admin)",
        "description": "Deletes a category from the Git repository. By default performs a hard delete (permanent removal). The 'hard' query parameter is available for explicit control but both soft and hard delete perform the same operation since categories don't have an isActive field. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the category",
            "example": "productivity-tools"
          },
          {
            "name": "hard",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Whether to perform hard delete (both soft and hard delete perform the same operation)",
            "example": true
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCategoryResponse"
                },
                "examples": {
                  "soft_delete": {
                    "summary": "Soft delete (same as hard delete)",
                    "value": {
                      "success": true,
                      "message": "Category deactivated successfully"
                    }
                  },
                  "hard_delete": {
                    "summary": "Hard delete",
                    "value": {
                      "success": true,
                      "message": "Category permanently deleted"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Unauthorized. Admin access required."
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Category with ID \"productivity-tools\" not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Session-based authentication using NextAuth.js"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "string",
            "description": "Additional error details"
          }
        },
        "required": ["success", "error"]
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique category identifier"
          },
          "name": {
            "type": "string",
            "description": "Category name"
          },
          "count": {
            "type": "integer",
            "description": "Number of items in this category (optional, only in some contexts)"
          },
          "isInactive": {
            "type": "boolean",
            "description": "Whether the category is inactive (optional, only in some contexts)"
          }
        },
        "required": ["id", "name"]
      },
...
      "GetCategoryResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/Category"
          }
        },
        "required": ["success", "data"]
      },
      "UpdateCategoryRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Updated category name",
            "example": "Enhanced Productivity Tools"
          }
        },
        "required": ["name"]
      },
      "UpdateCategoryResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/Category"
          },
          "message": {
            "type": "string",
            "example": "Category updated successfully"
          }
        },
        "required": ["success", "data", "message"]
      },
      "DeleteCategoryResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Category deleted successfully"
          }
        },
        "required": ["success", "message"]
      },
...
    },
    "responses": {
...
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Internal server error"
            }
          }
        }
      }
    }
  }
}